feat: Atlantis custom module source prefix list and CDKTF support
With the introduction of CDKTF (`*.tf.json`) and separation of Terraform module source paths,
Fogg Atlantis integration does not generate auto plan triggers for matching the new patterns.

Adding call to `os.Stat()` to detect if moduleSource is an existing path doesn't work due to:

1. plan/ci step runs before any changes are made in fs
1. moduleSource is relative to component nested path, complicating path checking logic

As an alternative solution, a new `module_prefixes` field was added to the Atlantis configuration,
defaulting to the original [`terraform/modules/`] prefix for backwards compatibility.

`*.tf.json` globbing pattern was also added to the fogg generated atlantis config file.